requirements.txt
requests==2.32.3
pandas==2.1.4
psycopg2-binary==2.9.10
apache-airflow-providers-postgres==5.13.1
